requirements.txt
playwright==1.46.0
openai==1.0.0
python-dotenv==1.0.0
pytest==8.3.2
pytest-html==3.1.1
pytest-xdist==3.6.1
Pillow==10.0.0
requests==2.31.0
fastapi==0.95.0
uvicorn==0.22.0
jinja2==3.1.2


aiofiles==23.1.0